+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 ### Fixed
+- Imgur images (up)loading ([#202](https://github.com/Tunous/Dawn/pull/202))
 
 ## [0.9.1] - 2020-05-05
 ### Fixed